Adding a due date selector
This commit adds a due date selector based on the simple one available
in RTM instead of a full date picker. Timezones are a pain. Will attempt
to fetch the users configured timezone from RTM and use that in the due
date chooser. If this is not taken in to account the user could select
tomorrow, based on the local timezone, and end up with the task having
a due date in two days time.

For example, a user from London has traveled to New York. It's a
Wednesday at 22:00 and they save a task with a due date of Friday.
Do they mean Friday New York time or Friday in London, it doesn't matter
as RTM will use London time as that is the user configured timezone. So,
they might expect it to be on Friday but it will show up in RTM with a
due date of Saturday.

Another edge case to consider, the user selects today. This is just
going to behave wrong in the example above, today is Thursday in London.

Fetching the users configured timezone allows the date picker to show
days based on their home location, not the local one.