Bumping to version 3.0
This commit updates the manifest with new keys. Makes the options page
look a little nicer in modern version of Firefox. The previous commit
removed a deprecated parameter from one of the browser calls which is
required for the extension to work on older versions. For this reason
the minimum support version is being bumped as well.